Extracted to future utils
